+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 
   - Major/Minor/Patch: This is an example changelog item, usually can be the commit message.
   - Append new items to make git merging easier.
+  - feat(Popover): implement `shouldClose` API
+  - fix(Checkbox): do not allow state change when `readOnly`
 </details>
 
 ## v0.70.0